+++ CampusConnect_Backend/src/main/java/com/campus_connect/CampusConnect_Backend/repository/ChatGroupRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.campus_connect.CampusConnect_Backend.models.ChatGroup;
 
-public interface ChatGroupRepository extends JpaRepository<ChatGroup, Long> {
+public interface ChatGroupRepository extends JpaRepository<ChatGroup, Integer> {
     Optional<ChatGroup> findByName(String name);
 }
